Add Java binaries endpoint to VanillaController
The VanillaController now includes a "GetJavaAsync" endpoint. This method will fetch and cache Java Runtime Environment (JRE) binaries, based on the given operating system, for an hour. In case the provided operating system is not supported, an error message that includes the list of supported platforms is returned. Additionally, error handling has been updated to return a proper "message" property, instead of an "error" property.