Update logging configuration and validation in Server Program
The logging configuration in the Server Program has been revised to be more robust and flexible. The minimum logging level has been changed from 'Information' to 'Verbose'. Separate logging levels are now defined for Debug and Release modes. In addition, the validation for the presence of log files has been updated from 'Any' check to checking the Length property.